Extends WooCommerce with Gravity Forms
Allows you to set a product as coming soon, instead of for sale. It adds
a checkbox in product admin to set a product as "coming soon". If
checked, it does not display a add to cart button but instead adds an
email form that says inform me when this product is ready. WARNING: this
requires the use of gravity forms, though it can be easily altered for
use with other forms plugins.